input: Moved window procedure and input to main
Moved window procedure and input to main.
